pyproject.toml
[project]
name = "markdown-to-googleslides"
version = "0.1.0"
description = "PowerPointファイルをGoogle Slidesにアップロード＆変換するCLIツール"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.174.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
]
